DDA-224: Extract active button from open in browser
